misc: fastrpc: Add early wakeup support for fastRPC
CPU wake up and context switch latency are significant in FastRPC
overhead for remote calls. As part of this change, DSP sends early
signal of completion to CPU and FastRPC driver detects early signal
on the given context and starts polling on a memory for actual
completion. Multiple different response flags are added to support
DSP user early hint of approximate time of completion, early response
from DSP user to wake up CPU and poll on memory for actual completion.
Complete signal is also added which is sent by DSP user in case of
timeout after early response is sent.

Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>
Link: https://lore.kernel.org/all/20240202064039.15505-10-quic_ekangupt@quicinc.com/
[mkrawczuk: backported from "[PATCH v1 09/16]" to v7.1-rc6
 (linux-msm/laptops-kernel). Adds enum fastrpc_response_flags; ctx fields
 early_wake_time/is_work_done/rsp_flags; channel cpuinfo_status/cpuinfo_todsp;
 poll_for_remote_response / fastrpc_wait_for_response /
 fastrpc_wait_for_completion; the USER_EARLY_SIGNAL state machine in
 fastrpc_internal_invoke replacing the simple wait_for_completion; rspv2
 handling in fastrpc_rpmsg_callback; and per-domain cpuinfo plumbing. Manual
 fixes: moved the poll/wait helpers ahead of fastrpc_internal_invoke; restored
 its bail block; relocated the cpuinfo_todsp lines into the proper
 switch(domain_id) cases in fastrpc_rpmsg_probe; restored the struct
 fastrpc_mmap_req_msg header; added #include <linux/delay.h> for udelay; added
 an interrupted=0 declaration.]
Signed-off-by: Mateusz Krawczuk <mat.krawczuk@gmail.com>